"Update visuals and features of solar system model"
The CSS, HTML, and JavaScript for the solar system model are updated, both for optimization and improved user experience. The size of the planets is scaled more accurately, orbits are added and can be toggled on and off, and a search bar is included for the user to search for specific planets which when selected, provides additional information and displayed in a separate card. Dark mode has also been improved to have better color contrast. The changes are made following user feedback and to improve the overall user interaction on the website.